+++ SBank/src/main/java/com/sbank/wrappers/WrapperATMAddMoneyToATM.java
@@ -9,16 +9,16 @@ public class WrapperATMAddMoneyToATM {
   public void setAmount(BigDecimal amount) {
     this.amount = amount;
   }
-  public BigDecimal getBankId() {
+  public Long getBankId() {
     return bankId;
   }
  
-  public void setBankId(BigDecimal bankId) {
+  public void setBankId(Long bankId) {
     this.bankId = bankId;
   }
 
   private BigDecimal amount;
-  private BigDecimal bankId;
+  private Long bankId;
   private Long atmID;
   
   public Long getAtmID() {
